fix: alias 'jx step create install values' -> 'jx step verify ingress'
to simplify the UX

fixes #5020
